+++ react/components/ReviewBlock.tsx
@@ -56,12 +56,12 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
     return Promise.all(
       data.map(async (item: any) => {
         const res: any = await checkRestriction(item.refid)
-        if (
-          res?.data?.productSuggestions?.products[0]?.items[0]?.itemId ===
-          item.sku
-        ) {
-          return item
-        }
+
+        const foundSku = res?.data?.productSuggestions?.products[0]?.items.find(
+          (suggestedItem) => suggestedItem.itemId === item.sku
+        )
+
+        return foundSku ? item : null
       })
     )
   }
@@ -87,6 +87,7 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
     'store/quickorder.available': messages.available,
     'store/quickorder.invalidPattern': messages.invalidPattern,
     'store/quickorder.skuNotFound': messages.skuNotFound,
+    'store/quickorder.partiallyAvailable': messages.partiallyAvailable,
     'store/quickorder.withoutStock': messages.withoutStock,
     'store/quickorder.withoutPriceFulfillment':
       messages.withoutPriceFulfillment,
@@ -129,7 +130,9 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
     refidData?.skuFromRefIds.items.forEach((item: any) => {
       if (!item.sellers) return
       item.sellers = item.sellers.filter(
-        (seller: any) => seller.availability === 'available'
+        (seller: any) =>
+          seller.availability === 'available' ||
+          seller.availability === 'partiallyAvailable'
       )
     })
 
@@ -165,13 +168,20 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
             return item != null
           })
         )
+
         restrictedData.forEach((item: any) => {
           mappedRefId[item.refid] = item
         })
       }
 
-      const errorMsg = (item: any) => {
+      const errorMsg = (item: any, sellerWithStock: string) => {
         let ret: any = null
+
+        /* order of precedence for errors
+         * 1) Item not found
+         * 2) Item availability
+         * 3) Item restriction
+         */
         const notfound = refIdNotFound.find((curr: any) => {
           return curr.refid === item.sku && curr.sku === null
         })
@@ -180,18 +190,37 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
           return curr.refid === item.sku && curr.sku !== null
         })
 
+        let selectedSellerHasPartialStock
         const foundHasStock =
           found?.sellers?.length &&
-          found.sellers.filter(
-            (seller: any) =>
-              seller.availability && seller.availability === 'available'
-          ).length
+          found.sellers.filter((seller: any) => {
+            if (seller.id === sellerWithStock) {
+              selectedSellerHasPartialStock =
+                seller.availability === 'partiallyAvailable'
+            }
+
+            return (
+              seller.availability &&
+              (seller.availability === 'available' ||
+                seller.availability === 'partiallyAvailable')
+            )
+          }).length
+
+        const itemRestricted = sellerWithStock
+          ? null
+          : `store/quickorder.limited`
+
+        const partialStockError = selectedSellerHasPartialStock
+          ? 'store/quickorder.partiallyAvailable'
+          : null
+
+        const availabilityError = foundHasStock
+          ? partialStockError
+          : `store/quickorder.withoutStock`
 
         ret = notfound
           ? 'store/quickorder.skuNotFound'
-          : foundHasStock
-          ? null
-          : `store/quickorder.withoutStock`
+          : availabilityError ?? itemRestricted
 
         return ret
       }
@@ -205,7 +234,9 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
           ? item.seller
           : item.sku && mappedRefId[item.sku]?.sellers?.length
           ? mappedRefId[item.sku]?.sellers.find(
-              (seller: any) => seller.availability === 'available'
+              (seller: any) =>
+                seller.availability === 'available' ||
+                seller.availability === 'partiallyAvailable'
             )?.id ?? ''
           : ''
 
@@ -216,6 +247,13 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
               )?.unitMultiplier ?? '1'
             : '1'
 
+        const sellerAvailableQuantity =
+          item.sku && mappedRefId[item.sku]?.sellers?.length
+            ? mappedRefId[item.sku]?.sellers.find(
+                (seller: any) => seller.id === sellerWithStock
+              )?.availableQuantity
+            : null
+
         return {
           ...item,
           sellers: item.sku ? mappedRefId[item.sku]?.sellers : [],
@@ -225,9 +263,8 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
           totalQuantity: sellerUnitMultiplier
             ? sellerUnitMultiplier * item.quantity
             : '',
-          error:
-            errorMsg(item) ??
-            (sellerWithStock ? null : `store/quickorder.limited`),
+          availableQuantity: sellerAvailableQuantity ?? item.quantity,
+          error: errorMsg(item, sellerWithStock),
         }
       })
 
@@ -262,9 +299,16 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
     onRefidLoading(true)
     const refids = _refids.length ? Array.from(new Set(_refids)) : []
 
+    const refIdQuantityMap = reviewed.reduce((prev, item) => {
+      return {
+        ...prev,
+        [item.sku]: item.quantity,
+      }
+    }, {})
+
     const query = {
       query: getRefIdTranslation,
-      variables: { refids, orderFormId, refIdSellerMap },
+      variables: { refids, orderFormId, refIdSellerMap, refIdQuantityMap },
     }
 
     try {
@@ -506,9 +550,15 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
         width: 75,
         cellRenderer: ({ cellData, rowData }: any) => {
           if (rowData.error) {
-            const text = intl.formatMessage(
-              errorMessage[cellData || 'store/quickorder.valid']
-            )
+            const errMsg = errorMessage[cellData || 'store/quickorder.valid']
+            const text =
+              errMsg === messages.partiallyAvailable
+                ? intl.formatMessage(errMsg, {
+                    quantity: rowData.availableQuantity,
+                    totalQuantity:
+                      rowData.availableQuantity * rowData.unitMultiplier,
+                  })
+                : intl.formatMessage(errMsg)
 
             return (
               <span className="pa3 br2 dib mr5 mv0">